# ui/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8001","description":"Generated server url"}],"paths":{"/api/persons/{id}":{"get":{"tags":["person-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonDetailResponseDTO"}}}}}},"put":{"tags":["person-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonModifyRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["person-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/persons":{"get":{"tags":["person-controller"],"operationId":"listUsers","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"email","in":"query","required":false,"schema":{"type":"string"}},{"name":"pageStart","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32"}},{"name":"pageSize","in":"query","required":false,"schema":{"maximum":200,"minimum":1,"type":"integer","format":"int32"}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","enum":["NAME","EMAIL"]}},{"name":"sortDesc","in":"query","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PagedResultResponseDTOPersonListItemResponseDTO"}}}}}},"post":{"tags":["person-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonModifyRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IdResponseDTOLong"}}}}}}},"/api/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}},"/api/auth/currentuser":{"get":{"tags":["auth-controller"],"operationId":"currentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}}},"components":{"schemas":{"PersonModifyRequestDTO":{"required":["address","email","firstName","lastName"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"address":{"type":"string"},"state":{"type":"string"},"phoneNumber":{"type":"string"}}},"IdResponseDTOLong":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"LogoutRequestDTO":{"type":"object","properties":{"username":{"type":"string"}}},"LoginRequestDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"UserResponseDTO":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"username":{"type":"string"}}},"PagedResultResponseDTOPersonListItemResponseDTO":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"elements":{"type":"array","items":{"$ref":"#/components/schemas/PersonListItemResponseDTO"}}}},"PersonListItemResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"}}},"PersonDetailResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"address":{"type":"string"},"state":{"type":"string"},"phoneNumber":{"type":"string"}}}}}}